changed aounts to int

+++ src/Person.java
@@ -2,12 +2,12 @@
 public class Person {
 	
 	private String firstName, lastName;
-	private int amountOfMoney;
+	private double amountOfMoney;
 	
 	public Person() {
 		firstName = "first";
 		lastName = "last";
-		amountOfMoney = 5000;
+		amountOfMoney = 5000.00;
 	}
 
 }
